requirements.txt
amqp==5.2.0
asgiref==3.8.1
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.30.2
azure-mgmt-core==1.4.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
Django==5.0.4
django-axes==6.5.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-modeltranslation==0.18.11
django-registration-redux==2.13
django-timezone-field==6.1.0
django_csp==3.8
gunicorn==22.0.0
idna==3.7
isodate==0.6.1
kombu==5.3.7
msrest==0.7.1
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
qrcode==7.4.2
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
stripe==9.3.0
svgwrite==1.4.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
